Order _sInit() invocations based on kit dependencies.
The following code was causing the SVM to crash (suppose Foo and Bar are in kit "akit")

class Foo extends Component {
  static inline Bar bar
}
final class Bar extends Component {
  void _iInit() { Sys.out.print("Bar initialized") }
}

The reason was that during ImageGen, Foo._sInit() was getting called before sys::Sys._sInit(). This was happening because static initialization was based on the inheritance ordering of the types (see OrderIrTypes.java). Inheritance ordering is based first on the inheritance hierarchy, and then on qname if two classes are at the same depth in the hierarchy.

This is not an appropriate ordering for static initialization - at the very least static initializers should be ordered based on kit dependencies. And that is what this fix does.