Manifest/Platform DBs load from sedonadev.org
The manifest and platform databases have been enhanced to try and download resources from sedonadev.org that are not present in the local databases.

We now support a property in sedona.properties for giving a list of failover websites to try.  http://sedonadev.org is the default:

sedonadev.org=<website1>,<website2>,...,<websiteN>