jsvm: bootstrap _sInit() invocations
The scode image ensures that all static initializers are invoked prior
to the main method.  The JSVM doesn't use an scode image, so now the
JSVM looks for a java class JsvmBootstrap in each kit named in the
app schema, and invokes a bootstrap() method on that class to cause
all static initializers for that kit to be executed.