FileStore c natives better implement File/FileStore API contract
This changeset contains enhancements to better implement the contract
required by FileStore/File/Instream/OutStream for return values. In
particular, we make sure -1 is returned when API requires it.

The "biggest" change was that sys_FileStore_doSize() has been
changed for POSIX and WIN32 so that they do not open the file
in order to get the size (behavior the FileStore API requires).

There are also some minor doc clarifications.