sox: close file involved in transfer when session ends abnormally
If the sox session ended abnormally, and a file was in transfer, it was not
getting closed. Now, when housekeeping closes a session that times out, it will
cause the file to be closed.  Hence, the file remains opened until the session
is forcefully closed by housekeeping!

Fixes Pacman Issue 14575: File left open when connection closes
  during file transfer