sedonac: -outDir works for all compilation targets
Issue 17524: -outDir option should work for compiling kits

Prior to this fix, -outDir was only honored when compiling kit documentation.
Now, -outDir will be honored when compiling kits, scode and apps.

When compiling a kit, the outDir will be treated as the KitDb root directory.
So, for example, if you are compiling the sys kit to outDir D:/temp, the kit
file will be save in D:\temp\sys\sys-<cksum>.kit

When compiling scode or app, the output file will be written directly to the
outDir.