sedonac: -outDir should work for all compilation targets
Issue 17524: -outDir option should work for compiling kits

This changest re-implements this fix. When compiling a kit with the -doc
and -outDir options we were deleting the kit when we genereated the docs. We
need to treat -outDir as basically sedona.home in this case. If the -outDir
option is specified, then output for compiling kits will go to

<outDir>/kits/<kitname>/kit-version.kit
<outDir>/doc/<kitname>/*.html

This re-implements changeset 500 (f95e3e583593)