dasp: backout maxSend implementation
This de-implements issue 17656: dasp session not closing after 3 failed retries
This de-implements changeset bb99b1b23936

Due to the aggressive nature of closing dasp sessions after 3 failed re-send
attempts, we are temporarily backing out this change.  This effectively
causes dasp to behave as it did prior to the original fix.

In the Java code, we have just commented out the relevant code (it was a small
change). In the Sedona code, we have backed out all the code so as to not
increase the scode size of the sox kit.