sedonac: deprecate 'name' attribute from sedonaCode xml
Issue 17693: scode xml name attribute ignored

The name attribute of sedonaCode xml was completely ignored by the compiler
and did not impact the name of the scode image that got generated. Therefore,
we are deprecating this attribute and scode images will simply continue to
have the same name as the sedonaCode xml input file, but with a ".scode"
extension.

The compiler now generates a warning when this attribute is present, and it
is no longer required.