sedonac: change how source files are included a kit
Issue 16866: include kit.xml if includeSource is true

Currently, when includeSource=true, the kit.xml is not included in the kit's
source/ directory.  In order to support this enhancement the following changes
were made:

1) Do not allow source directories in kit.xml that are not rooted in the
same directory as the kit.xml. This is a compiler error now.

2) Maintain source directory tree when writing source into the kit file.  Prior
to this fix, all the source was flattened into the source/ directory. In order
to keep the input kit.xml "as-is" we must maintain the directory structure
to the source files.

3) Do not include source from types that were included with <include> directives
in the kit.xml.  In order to keep the kit.xml "as-is", it no longer made
sense to include the source for included types.