sedona: better error checking for listing kit parts
Issue 17969: ManifestDb shot not list KitParts that have wrong name

Prior to this fix, if a manifest was in the wrong directory, for example:
foo-aabbccdd.xml was accidentally in the manifests/bar directory, it would
still return this as a kit part during listing.  Now we check to make
sure the file name starts with the kit being listed.