inet: TcpSocket does not close when finishConnect() fails
Issue 17651: TcpSocket does not close when finishConnect() fails

In the case where the state of the socket is "Failed" after a call to
finishConnect(), then a subsequent call to close() would not actually
cause the underlying socket to be closed.

This changeset fixes that issue and updates the documentation to make
it explicitly clear that close() still must be called when finishConnect()
fails.