inet: load IpAddr from Buf or byte[]
Issue 18296: enhance IpAddr to be built with raw Buf

This enhancement provides a way to load an IpAddr with a raw byte[]
representation of the ip address.  A convenience method is also provided
to load an IpAddr with a Buf containing the raw bytes for the ip address.

The implementation requires the byte[] to be either 4-bytes (IPv4) or
16-bytes (IPv6), otherwise the ip address will not be loaded.