sedonac: code fixes based on FindBugs analysis
Made a number of minor changes to handle bugs reported by the FindBugs
plugin for Eclipse.  Most fixes are related to supplying hashCode() or
equals() to classes that don't implement them (when they should).

Other types of fixes are removing unused fields or local variables (only
in certain cases).  Fixing imports to be explicit.

See diffs for full set of changes.